requirements.txt
tensorflow-cpu==2.8.0
Flask==2.0.3
flask-cors==3.0.10
numpy==1.19.5
pillow==8.4.0
gunicorn==20.1.0
